esp32 compiler is a bit picky about max() definition

+++ src/FastAccelStepper.cpp
@@ -588,7 +588,7 @@ void FastAccelStepper::setDelayToDisable(uint16_t delay_ms) {
     // ensure minimum time
     delay_count = 2;
   }
-  _off_delay_count = max(delay_count, 1);
+  _off_delay_count = max(delay_count, (uint16_t)1);
 }
 int8_t FastAccelStepper::moveTo(int32_t position) {
   return _rg.moveTo(position, &fas_queue[_queue_num].queue_end);
